start_screen_di some fixs
