+++ buildSrc/src/main/java/Dependencies.kt
@@ -28,6 +28,7 @@ object Versions {
     //Design
     const val appcompat = "1.5.1"
     const val material = "1.7.0"
+    const val swipe = "1.1.0"
 
     //Kotlin
     const val core = "1.8.0"
@@ -65,6 +66,7 @@ object Versions {
 object Design {
     const val appcompat = "androidx.appcompat:appcompat:${Versions.appcompat}"
     const val material = "com.google.android.material:material:${Versions.material}"
+    const val swipe = "androidx.swiperefreshlayout:swiperefreshlayout:${Versions.swipe}"
 }
 
 object Kotlin {
