+++ featureStartScreen/build.gradle.kts
@@ -23,7 +23,7 @@ dependencies {
 
     //AndroidX
     implementation(Design.appcompat)
-    implementation("androidx.swiperefreshlayout:swiperefreshlayout:1.1.0")
+    implementation(Design.swipe)
 
 
     //Design
