+++ featureStartScreen/src/main/java/com/bestscore/featurestartscreen/StartScreenFragment.kt
@@ -1,7 +1,6 @@
 package com.bestscore.featurestartscreen
 
 import android.content.Context
-import android.graphics.Canvas
 import android.os.Bundle
 import android.view.View
 import androidx.core.view.isVisible
@@ -12,14 +11,10 @@ import androidx.lifecycle.get
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
-import androidx.recyclerview.widget.RecyclerView.ItemDecoration
 import by.kirich1409.viewbindingdelegate.viewBinding
 import com.bestscore.core.navigation.navigate
 import com.bestscore.featurestartscreen.databinding.FragmentStartScreenBinding
 import com.bestscore.featurestartscreen.di.StartScreenComponentViewModel
-import com.bestscore.featurestartscreen.swipe.SwipeController
-import com.bestscore.featurestartscreen.swipe.SwipeControllerActions
 import com.bestscore.utils.makeToast
 import javax.inject.Inject
 
@@ -46,8 +41,6 @@ class StartScreenFragment : Fragment(R.layout.fragment_start_screen) {
         )
     }
 
-    //private val swipeController: SwipeController = SwipeController()
-    // private val itemTouchHelper: ItemTouchHelper = ItemTouchHelper(swipeController)
 
     override fun onAttach(context: Context) {
         ViewModelProvider(this)
@@ -89,131 +82,14 @@ class StartScreenFragment : Fragment(R.layout.fragment_start_screen) {
             }
             adapter = this@StartScreenFragment.adapter
             layoutManager = LinearLayoutManager(requireContext())
+            val swipeController = SwipeController(requireContext())
+            val itemTouchHelper = ItemTouchHelper(swipeController)
+            itemTouchHelper.attachToRecyclerView(this)
 
-            setItemTouchHelper()
+            //  setItemTouchHelper()
         }
 
     }
-
-    private fun setItemTouchHelper() {
-
-        ItemTouchHelper(object : ItemTouchHelper.Callback() {
-            //limit to swipe same as buttons
-            private val limitScrollX = dpToPx(96f, this@StartScreenFragment.requireContext())
-            /* private val limitScrollX =
-                 this@StartScreenFragment.context?.let {
-                     dpToPx(
-                         48f,
-                         it
-                     )
-                 } */
-
-            private var currentScrollX = 0
-            private var currentScrollXWhenInActive = 0
-            private var initXWhenInActive = 0f
-            private var firstInActive = false
-
-
-            override fun getMovementFlags(
-                recyclerView: RecyclerView,
-                viewHolder: RecyclerView.ViewHolder
-            ): Int {
-                val dragFlags = 0
-                val swipeFlags = ItemTouchHelper.LEFT or ItemTouchHelper.RIGHT
-                return makeMovementFlags(dragFlags, swipeFlags)
-            }
-
-            override fun onMove(
-                recyclerView: RecyclerView,
-                viewHolder: RecyclerView.ViewHolder,
-                target: RecyclerView.ViewHolder
-            ): Boolean {
-                return true
-            }
-
-            override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {}
-
-            override fun getSwipeThreshold(viewHolder: RecyclerView.ViewHolder): Float {
-                return Integer.MAX_VALUE.toFloat()
-            }
-
-            override fun getSwipeEscapeVelocity(defaultValue: Float): Float {
-                return Integer.MAX_VALUE.toFloat()
-            }
-
-            override fun onChildDraw(
-                c: Canvas,
-                recyclerView: RecyclerView,
-                viewHolder: RecyclerView.ViewHolder,
-                dX: Float,
-                dY: Float,
-                actionState: Int,
-                isCurrentlyActive: Boolean
-            ) {
-                if (actionState == ItemTouchHelper.ACTION_STATE_SWIPE) {
-
-                    if (dX == 0f) {
-                        currentScrollX = viewHolder.itemView.scrollX
-                        firstInActive = true
-                    }
-
-                    if (isCurrentlyActive) {
-                        //swipe with finger
-
-                        var scrollOffset = currentScrollX + (-dX).toInt()
-                        if (scrollOffset > limitScrollX) {
-                            scrollOffset = limitScrollX
-                        } else if (scrollOffset < 0) {
-                            scrollOffset = 0
-                        }
-                        viewHolder.itemView.scrollTo(scrollOffset, 0)
-                    } else {
-
-                        //swipe with auto animation
-                        if (firstInActive) {
-                            firstInActive = false
-                            currentScrollXWhenInActive = viewHolder.itemView.scrollX
-                            initXWhenInActive = dX
-                        }
-
-                        if (viewHolder.itemView.scrollX < limitScrollX) {
-                            viewHolder.itemView.scrollTo(
-                                (currentScrollXWhenInActive * dX / initXWhenInActive).toInt(),
-                                0
-                            )
-                        }
-
-                    }
-
-                }
-            }
-
-            override fun clearView(
-                recyclerView: RecyclerView,
-                viewHolder: RecyclerView.ViewHolder
-            ) {
-                super.clearView(recyclerView, viewHolder)
-
-                if (viewHolder.itemView.scrollX > limitScrollX) {
-                    viewHolder.itemView.scrollTo(limitScrollX,0)
-                }
-                else if (viewHolder.itemView.scrollX < 0) {
-                    viewHolder.itemView.scrollTo(0,0)
-                }
-
-
-            }
-
-        }).apply {
-            attachToRecyclerView(binding.rvLastTemplates)
-        }
-
-    }
-
-    private fun dpToPx(dpValue: Float, context: Context): Int {
-        return (dpValue * context.resources.displayMetrics.density).toInt()
-    }
-
 
     private fun collectFlow() {
         viewLifecycleOwner.lifecycleScope.launchWhenStarted {
