--- featureStartScreen/src/main/java/com/bestscore/featurestartscreen/swipe/ButtonsState.kt
@@ -1,5 +1,0 @@
-package com.bestscore.featurestartscreen.swipe
-
-internal enum class ButtonsState {
-    GONE, LEFT_VISIBLE, RIGHT_VISIBLE
-}
