--- featureStartScreen/src/main/java/com/bestscore/featurestartscreen/swipe/SwipeControllerActions.kt
@@ -1,11 +1,0 @@
-package com.bestscore.featurestartscreen.swipe
-
-import android.util.Log
-
-abstract class SwipeControllerActions {
-
-    open fun onRightClicked(position: Int) {
-
-    }
-
-}
